requirements.txt
Flask==2.0.3
Pillow==8.4.0
